moved FromBridgedChainMessagesProof to bp-messages
