+++ bin/millau/runtime/src/rialto_messages.rs
@@ -50,9 +50,6 @@ pub type ToRialtoMessagePayload = messages::source::FromThisChainMessagePayload;
 /// Message payload for Rialto -> Millau messages.
 pub type FromRialtoMessagePayload = messages::target::FromBridgedChainMessagePayload;
 
-/// Messages proof for Rialto -> Millau messages.
-pub type FromRialtoMessagesProof = messages::target::FromBridgedChainMessagesProof<bp_rialto::Hash>;
-
 /// Call-dispatch based message dispatch for Rialto -> Millau messages.
 pub type FromRialtoMessageDispatch =
 	bridge_runtime_common::messages_xcm_extension::XcmBlobMessageDispatch<
