+++ bin/runtime-common/src/refund_relayer_extension.rs
@@ -647,7 +647,6 @@ where
 mod tests {
 	use super::*;
 	use crate::{
-		messages::target::FromBridgedChainMessagesProof,
 		messages_call_ext::{
 			BaseMessagesProofInfo, ReceiveMessagesDeliveryProofInfo, ReceiveMessagesProofInfo,
 			UnrewardedRelayerOccupation,
@@ -655,8 +654,8 @@ mod tests {
 		mock::*,
 	};
 	use bp_messages::{
-		source_chain::FromBridgedChainMessagesDeliveryProof, InboundLaneData, MessageNonce,
-		OutboundLaneData, UnrewardedRelayersState,
+		source_chain::FromBridgedChainMessagesDeliveryProof, target_chain::FromBridgedChainMessagesProof,
+		InboundLaneData, MessageNonce, OutboundLaneData, UnrewardedRelayersState,
 	};
 	use bp_parachains::{BestParaHeadHash, ParaInfo};
 	use bp_polkadot_core::parachains::{ParaHeadsProof, ParaId};
