+++ tools/runtime-codegen/src/main.rs
@@ -141,7 +141,7 @@ fn main() -> color_eyre::Result<()> {
 			TypeSubstitute::simple("bp_polkadot_core::parachains::ParaId"),
 			TypeSubstitute::simple("bp_polkadot_core::parachains::ParaHeadsProof"),
 			TypeSubstitute::simple(
-				"bridge_runtime_common::messages::target::FromBridgedChainMessagesProof",
+				"bp_messages::target_chain::FromBridgedChainMessagesProof",
 			),
 			TypeSubstitute::simple(
 				"bp_messages::source_chain::FromBridgedChainMessagesDeliveryProof",
